Function backend_rust::db::elections::create_election

pub fn create_election(
    election: &ElectionInput, 
    creator_id: &str, 
    conn: &DbConnection
) -> Result<(Election, Vec<Choice>), Error>

Create a new election in the database

This inserts a new election given an election input and the ID of the user creating it. This will error if the insert fails for some reason (at this point it should only be due to connection errors).

Arguments

election - The input object with the user input data for the new election creator_id - The id of the user calling the creation function conn - The database connection to execute the insert on

Example

//currently not tested, since it's annoying having to set up a test DB
let election_input = ElectionInput {
    name: "test_name".to_string(),
    description: Some("This is a test description".to_string()),
    choices: Some(vec!["test1".to_string(), "test2".to_string()]),
    permissions: None,
    start_date: None,
    end_date: None,
    importance: None
};

let result = elections::create_election(&election_input, "auth0|asdqwe", &conn);